initial start throws wrapped ioexception, filter chain selector exception
