+++ xds/build.gradle
@@ -99,7 +99,6 @@ javadoc {
     exclude 'io/grpc/xds/Envoy*'
     exclude 'io/grpc/xds/TlsContextManager.java'
     exclude 'io/grpc/xds/XdsAttributes.java'
-    exclude 'io/grpc/xds/XdsClientWrapperForServerSds.java'
     exclude 'io/grpc/xds/XdsInitializationException.java'
     exclude 'io/grpc/xds/XdsNameResolverProvider.java'
     exclude 'io/grpc/xds/internal/**'
